feat: Setup basic home screen
This commit sets up a basic home screen using Jetpack Compose. It includes a `LazyColumn` to display a list of animal cards.

The `HomeScreen` composable now accepts a list of `Animals` and a `NavController` as parameters.

A preview for the `HomeScreen` was removed as it requires parameters that are not easily provided in a simple preview.